feat: плавное переключение тогла
