+++ package.json
@@ -19,6 +19,8 @@
     "editorconfig": "^2.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.2",
+    "react-icons": "^4.11.0",
     "react-router-dom": "^6.18.0",
     "sass": "^1.69.5"
   },
