+++ src/components/AuthPopup/AuthPopup.tsx
@@ -1,27 +1,15 @@
 
-import { FC, useEffect, useState } from 'react';
+import { FC, useState } from 'react';
 import { Popup } from '../Popup/Popup';
 import { ButtonForAuth } from '../UI/ButtonForAuth/ButtonForAuth';
-import { Checkbox } from '../UI/Checkbox/Checkbox';
-import { TEXT_FOR_AUTH_CHECKBOX } from '../../utils/constants';
 import { Input } from '../UI/Input/Input';
 import { InputTypes } from '../UI/Input/InputTypes';
 import { EMAIL_VALIDATION_CONFIG, PASSWORD_VALIDATION_CONFIG } from '../../utils/constants';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { ISignInData, ISignInFields } from './PopupForAuthTypes';
-import { IShippingFields } from './PopupForAuthTypes';
-import { ToggleButton } from '../UI/ToggleButton/ToggleButton';
+import { ISignInFields } from './AuthPopupTypes';
+import { IShippingFields } from './AuthPopupTypes';
 
-
-// interface IPopupForAuth {
-// 	isOpened: boolean;
-// 	// setIsOpened: React.Dispatch<React.SetStateAction<boolean>>;
-// }
-
-export const PopupForAuth: FC = ({
-	// isOpened,
-	// setIsOpened,
-}) => {
+export const PopupForAuth: FC = () => {
 	const [authError, setAuthError] = useState(false);
 	const {
 		register,
@@ -35,23 +23,10 @@ export const PopupForAuth: FC = ({
 		console.log(data);
 		reset();
 	};
-
-
-
-	// useEffect(() => {
-	// 	reset();
-	// 	setAuthError(false);
-	// }, []);
 	return (
-		<Popup
-		// setIsOpened={openedLoginPopup}
-		// closePopup={(param: boolean) => dispatch(openLoginPopup(param))}
-		>
-			<form
-				onSubmit={handleSubmit(onSubmit)}
-			>
-			<ToggleButton itsLoginPopup={true}/>
-			<Input
+		<Popup>
+			<form onSubmit={handleSubmit(onSubmit)}>
+				<Input
 						inputType={InputTypes.email}
 						labelText='e-mail'
 						validation={{
@@ -71,12 +46,7 @@ export const PopupForAuth: FC = ({
 							Неверный логин или пароль.
 						</p>
 					) : null}
-					{/* {TEXT_FOR_AUTH_CHECKBOX.map((i)=>{
-				return (
-				<Checkbox label={i.text} key={i.id} selected={false}/>
-				)
-			})} */}
-			<ButtonForAuth isValid={isValid} itsLoginPopup={true} title='Войти'/>
+				<ButtonForAuth isValid={isValid} itsLoginPopup={true} title='Войти'/>
 			</form>
 		</Popup>
 	);
