+++ src/components/Popup/PopupTypes.tsx
@@ -1,8 +1,5 @@
 import { ReactNode } from 'react';
-
 export interface IPopup {
 	children: ReactNode;
 	setIsOpened?: boolean;
-	// closePopup(param: boolean)?: void;
-	// setIsOpened: React.Dispatch<React.SetStateAction<boolean>>;
 }
