+++ src/components/RegPopup/RegPopup.tsx
@@ -8,8 +8,8 @@ import { Input } from '../UI/Input/Input';
 import { InputTypes } from '../UI/Input/InputTypes';
 import { EMAIL_VALIDATION_CONFIG, PASSWORD_VALIDATION_CONFIG, INN_VALIDATION_CONFIG, ORGNAME_VALIDATION_CONFIG, VALIDATION_SETTINGS, NAME_VALIDATION_CONFIG } from '../../utils/constants';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import {  ISignUpFields, ISignUpData } from '../AuthPopup/PopupForAuthTypes';
-import { IShippingFields } from '../AuthPopup/PopupForAuthTypes';
+import {  ISignUpFields, ISignUpData } from '../AuthPopup/AuthPopupTypes';
+import { IShippingFields } from '../AuthPopup/AuthPopupTypes';
 import { ToggleButton } from '../UI/ToggleButton/ToggleButton';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../services/redux/store';
