+++ src/components/UI/ToggleButton/ToggleButton.tsx
@@ -6,20 +6,22 @@ import { checkBoxState } from './ToggleButtonSlice';
 import {  useSelector } from 'react-redux';
 import { RootState } from '../../../services/redux/store';
 
-export const ToggleButton: FC<IToggleButton> = ({ label }) => {
+export const ToggleButton: FC<IToggleButton> = () => {
 		const dispatch = useDispatch();
-		const [checked, setChecked] = useState((useSelector((state: RootState) => state.toggleBtn.value))|| false);
+		const [checked, setChecked] = useState((useSelector((state: RootState) => state.toggleBtn.value)));
 		function handleClick(){
+			console.log(!checked)
+			dispatch(checkBoxState(!checked))
 			setChecked(!checked)
 		}
-		useEffect(()=>{
-			console.log(checked)
-			dispatch(checkBoxState(checked))
-		}, [checked])
+
 	return (
-		<label className={styles.wrapper} htmlFor="toggleSwitch" >
-			<input type="checkbox" className={styles.toggle} id="toggleSwitch" checked={checked} onChange={handleClick}/>
-			<p className={styles.label}>{label ?? ''}</p>
+		<div className={styles.wrapper}>
+		<input type="checkbox" className={styles.toggle} id="toggleSwitch" checked={checked} onChange={()=>handleClick()}/>
+		<label className={styles.textlabel} htmlFor="toggleSwitch" >
+				<p className={styles.text}>erfg</p>
+				<p className={styles.text}>erfg</p>
 		</label>
+		</div>
 	);
 };
