+++ src/components/UI/ToggleButton/ToggleButtonTypes.tsx
@@ -1,6 +1,3 @@
 export interface IToggleButton {
-	buttonText?: string;
-	itsLoginPopup?: boolean;
-	label?: string;
 	value?: boolean;
 }
